Rust 2022.23 i32->i16 (.51s->.43s)

+++ rust/2022/d23/src/main.rs
@@ -1,8 +1,8 @@
 use std::collections::HashSet;
 
-const DIRS: [(i32, i32); 7] = [(-1,0), (1,0), (0,-1), (0,1), (-1,0), (1,0), (0,-1)];
-const D8: [(i32, i32); 8] = [(0, 1), (0, -1), (-1, 0), (1, 0), (1, 1), (1, -1), (-1, 1), (-1, -1)];
-type Elf = (i32, i32);
+const DIRS: [(i16, i16); 7] = [(-1,0), (1,0), (0,-1), (0,1), (-1,0), (1,0), (0,-1)];
+const D8: [(i16, i16); 8] = [(0, 1), (0, -1), (-1, 0), (1, 0), (1, 1), (1, -1), (-1, 1), (-1, -1)];
+type Elf = (i16, i16);
 fn dest(elves: &HashSet<Elf>, elf: Elf, t: usize) -> Elf {
     if D8.iter().all(|&(di, dj)| !elves.contains(&(di+elf.0, dj+elf.1))) { return elf; }
     for &(di, dj) in DIRS[t%4..t%4+4].iter() {
@@ -16,7 +16,7 @@ fn dest(elves: &HashSet<Elf>, elf: Elf, t: usize) -> Elf {
 fn main() {
     let s = std::fs::read_to_string("../../../data/2022/23.input.txt").unwrap();
     let elves: HashSet<_> = s.lines().enumerate().flat_map(|(i, l)| 
-        l.chars().enumerate().filter(|&(_j, c)| c == '#').map(move |(j, _c)| (i as i32, j as i32))).collect();
+        l.chars().enumerate().filter(|&(_j, c)| c == '#').map(move |(j, _c)| (i as i16, j as i16))).collect();
     {
         let mut elves = elves.clone();
         for i in 0..10 {
@@ -26,7 +26,7 @@ fn main() {
         let imx = elves.iter().map(|e| e.0).max().unwrap();
         let jmn = elves.iter().map(|e| e.1).min().unwrap();
         let jmx = elves.iter().map(|e| e.1).max().unwrap();
-        println!("{}", (imx-imn+1)*(jmx-jmn+1)-elves.len() as i32);
+        println!("{}", (imx-imn+1)*(jmx-jmn+1)-elves.len() as i16);
     }
     
     {
@@ -41,7 +41,7 @@ fn main() {
     }
 }
 
-fn step(elves: &mut HashSet<(i32, i32)>, i: usize) -> HashSet<(i32, i32)> {
+fn step(elves: &mut HashSet<(i16, i16)>, i: usize) -> HashSet<(i16, i16)> {
     let prev = elves.clone();
     elves.clear();
     for &e in &prev {
